separate base.ts to gitlab-base.ts and redmine-base.ts
